+++ packages/amplify-category-auth/src/__tests__/provider-utils/awscloudformation/auth-stack-builder/auth-stack-transform.test.ts
@@ -302,12 +302,12 @@ describe('Check Auth Template', () => {
     const authTransform = new AmplifyAuthTransform(resourceName);
     const mock_template = await authTransform.transform(context_stub_typed);
     expect(mock_template.Resources?.UserPoolClient.Properties.AllowedOAuthFlows).toMatchInlineSnapshot(`
-      Array [
+      [
         "code",
       ]
     `);
     expect(mock_template.Resources?.UserPoolClient.Properties.AllowedOAuthScopes).toMatchInlineSnapshot(`
-      Array [
+      [
         "phone",
         "email",
         "openid",
@@ -316,19 +316,19 @@ describe('Check Auth Template', () => {
       ]
     `);
     expect(mock_template.Resources?.UserPoolClient.Properties.CallbackURLs).toMatchInlineSnapshot(`
-      Array [
+      [
         "https://localhost:3000/",
       ]
     `);
     expect(mock_template.Resources?.UserPoolClient.Properties.LogoutURLs).toMatchInlineSnapshot(`undefined`);
 
     expect(mock_template.Resources?.UserPoolClientWeb.Properties.AllowedOAuthFlows).toMatchInlineSnapshot(`
-      Array [
+      [
         "code",
       ]
     `);
     expect(mock_template.Resources?.UserPoolClientWeb.Properties.AllowedOAuthScopes).toMatchInlineSnapshot(`
-      Array [
+      [
         "phone",
         "email",
         "openid",
@@ -337,7 +337,7 @@ describe('Check Auth Template', () => {
       ]
     `);
     expect(mock_template.Resources?.UserPoolClientWeb.Properties.CallbackURLs).toMatchInlineSnapshot(`
-      Array [
+      [
         "https://localhost:3000/",
       ]
     `);
@@ -364,7 +364,7 @@ describe('Check Auth Template', () => {
         NoEcho: true,
       },
       `
-      Object {
+      {
         "NoEcho": true,
         "Type": "String",
       }
@@ -376,7 +376,7 @@ describe('Check Auth Template', () => {
         NoEcho: true,
       },
       `
-      Object {
+      {
         "NoEcho": true,
         "Type": "String",
       }
@@ -389,7 +389,7 @@ describe('Check Auth Template', () => {
         NoEcho: true,
       },
       `
-      Object {
+      {
         "NoEcho": true,
         "Type": "String",
       }
@@ -401,7 +401,7 @@ describe('Check Auth Template', () => {
         NoEcho: true,
       },
       `
-      Object {
+      {
         "NoEcho": true,
         "Type": "String",
       }
@@ -414,7 +414,7 @@ describe('Check Auth Template', () => {
         NoEcho: true,
       },
       `
-      Object {
+      {
         "NoEcho": true,
         "Type": "String",
       }
@@ -426,7 +426,7 @@ describe('Check Auth Template', () => {
         NoEcho: true,
       },
       `
-      Object {
+      {
         "NoEcho": true,
         "Type": "String",
       }
@@ -439,7 +439,7 @@ describe('Check Auth Template', () => {
         NoEcho: true,
       },
       `
-      Object {
+      {
         "NoEcho": true,
         "Type": "String",
       }
@@ -451,7 +451,7 @@ describe('Check Auth Template', () => {
         NoEcho: true,
       },
       `
-      Object {
+      {
         "NoEcho": true,
         "Type": "String",
       }
@@ -463,7 +463,7 @@ describe('Check Auth Template', () => {
         NoEcho: true,
       },
       `
-      Object {
+      {
         "NoEcho": true,
         "Type": "String",
       }
@@ -488,7 +488,7 @@ describe('Check Auth Template', () => {
     expect(mock_template.Resources!.HostedUIFacebookProviderResource.Properties.AttributeMapping).toMatchInlineSnapshot(
       { email: 'email', username: 'id' },
       `
-      Object {
+      {
         "email": "email",
         "username": "id",
       }
@@ -501,14 +501,14 @@ describe('Check Auth Template', () => {
         client_secret: { Ref: 'facebookAppSecretUserPool' },
       },
       `
-      Object {
-        "authorize_scopes": Object {
+      {
+        "authorize_scopes": {
           "Ref": "facebookAuthorizeScopes",
         },
-        "client_id": Object {
+        "client_id": {
           "Ref": "facebookAppIdUserPool",
         },
-        "client_secret": Object {
+        "client_secret": {
           "Ref": "facebookAppSecretUserPool",
         },
       }
@@ -518,7 +518,7 @@ describe('Check Auth Template', () => {
     expect(mock_template.Resources!.HostedUIGoogleProviderResource.Properties.AttributeMapping).toMatchInlineSnapshot(
       { email: 'email', username: 'sub' },
       `
-      Object {
+      {
         "email": "email",
         "username": "sub",
       }
@@ -531,14 +531,14 @@ describe('Check Auth Template', () => {
         client_secret: { Ref: 'googleAppSecretUserPool' },
       },
       `
-      Object {
-        "authorize_scopes": Object {
+      {
+        "authorize_scopes": {
           "Ref": "googleAuthorizeScopes",
         },
-        "client_id": Object {
+        "client_id": {
           "Ref": "googleAppIdUserPool",
         },
-        "client_secret": Object {
+        "client_secret": {
           "Ref": "googleAppSecretUserPool",
         },
       }
@@ -548,7 +548,7 @@ describe('Check Auth Template', () => {
     expect(mock_template.Resources!.HostedUILoginWithAmazonProviderResource.Properties.AttributeMapping).toMatchInlineSnapshot(
       { email: 'email', username: 'user_id' },
       `
-      Object {
+      {
         "email": "email",
         "username": "user_id",
       }
@@ -561,14 +561,14 @@ describe('Check Auth Template', () => {
         client_secret: { Ref: 'loginwithamazonAppSecretUserPool' },
       },
       `
-      Object {
-        "authorize_scopes": Object {
+      {
+        "authorize_scopes": {
           "Ref": "loginwithamazonAuthorizeScopes",
         },
-        "client_id": Object {
+        "client_id": {
           "Ref": "loginwithamazonAppIdUserPool",
         },
-        "client_secret": Object {
+        "client_secret": {
           "Ref": "loginwithamazonAppSecretUserPool",
         },
       }
@@ -578,7 +578,7 @@ describe('Check Auth Template', () => {
     expect(mock_template.Resources!.HostedUISignInWithAppleProviderResource.Properties.AttributeMapping).toMatchInlineSnapshot(
       { email: 'email' },
       `
-      Object {
+      {
         "email": "email",
       }
     `,
@@ -592,20 +592,20 @@ describe('Check Auth Template', () => {
         team_id: { Ref: 'signinwithappleTeamIdUserPool' },
       },
       `
-      Object {
-        "authorize_scopes": Object {
+      {
+        "authorize_scopes": {
           "Ref": "signinwithappleAuthorizeScopes",
         },
-        "client_id": Object {
+        "client_id": {
           "Ref": "signinwithappleClientIdUserPool",
         },
-        "key_id": Object {
+        "key_id": {
           "Ref": "signinwithappleKeyIdUserPool",
         },
-        "private_key": Object {
+        "private_key": {
           "Ref": "signinwithapplePrivateKeyUserPool",
         },
-        "team_id": Object {
+        "team_id": {
           "Ref": "signinwithappleTeamIdUserPool",
         },
       }