+++ packages/amplify-e2e-tests/src/__tests__/parameter-store-1.test.ts
@@ -9,16 +9,22 @@ import {
   getAppId,
   getProjectMeta,
   initJSProjectWithProfile,
+  removeFunction,
 } from '@aws-amplify/amplify-e2e-core';
 import { addEnvironmentCarryOverEnvVars, checkoutEnvironment, removeEnvironment } from '../environment/env';
 
 describe('upload and delete parameters', () => {
   let projRoot: string;
+  let deletionRequired = false;
   beforeEach(async () => {
     projRoot = await createNewProjectDir('upload-delete-parameters-test');
   });
 
   afterEach(async () => {
+    // In case test failed after init, but before delete
+    if (deletionRequired) {
+      await deleteProject(projRoot);
+    }
     deleteProjectDir(projRoot);
   });
 
@@ -28,7 +34,7 @@ describe('upload and delete parameters', () => {
     const envVariableName = 'envVariableName';
     const envVariableValue = 'envVariableValue';
     await initJSProjectWithProfile(projRoot, { disableAmplifyAppCreation: false, envName: firstEnvName });
-
+    deletionRequired = true;
     const meta = getProjectMeta(projRoot);
     expect(meta).toBeDefined();
     const appId = getAppId(projRoot);
@@ -37,6 +43,8 @@ describe('upload and delete parameters', () => {
     expect(region).toBeDefined();
 
     const fnName = `parameterstestfn${generateRandomShortId()}`;
+    const fnName2 = `parameterstestfn${generateRandomShortId()}`;
+
     await addFunction(
       projRoot,
       {
@@ -49,23 +57,32 @@ describe('upload and delete parameters', () => {
       },
       'nodejs',
     );
+
     await amplifyPushAuth(projRoot);
-    const expectedParamsAfterAddFunc = [
-      { name: 'deploymentBucketName' },
-      { name: envVariableName, value: envVariableValue },
-      { name: 's3Key' },
-    ];
+    const deploymentBucketAndKey = [{ name: 'deploymentBucketName' }, { name: 's3Key' }];
+    const expectedParamsAfterAddFunc = [...deploymentBucketAndKey, { name: envVariableName, value: envVariableValue }];
     await expectParametersOptionalValue(expectedParamsAfterAddFunc, [], region, appId, firstEnvName, 'function', fnName);
 
     await addEnvironmentCarryOverEnvVars(projRoot, { envName: secondEnvName });
+    await addFunction(projRoot, { name: fnName2, functionTemplate: 'Hello World' }, 'nodejs');
     await amplifyPushAuth(projRoot);
-    const expectedParamsAfterAddEnv = [
-      { name: 'deploymentBucketName' },
-      { name: envVariableName, value: envVariableValue },
-      { name: 's3Key' },
-    ];
+    const expectedParamsAfterAddEnv = [...deploymentBucketAndKey, { name: envVariableName, value: envVariableValue }];
     await expectParametersOptionalValue(expectedParamsAfterAddFunc, [], region, appId, firstEnvName, 'function', fnName);
     await expectParametersOptionalValue(expectedParamsAfterAddEnv, [], region, appId, secondEnvName, 'function', fnName);
+    await expectParametersOptionalValue(deploymentBucketAndKey, [], region, appId, secondEnvName, 'function', fnName2);
+
+    await removeFunction(projRoot, fnName2);
+    await amplifyPushAuth(projRoot);
+
+    await expectParametersOptionalValue(
+      [],
+      deploymentBucketAndKey.map((pair) => pair.name),
+      region,
+      appId,
+      secondEnvName,
+      'function',
+      fnName2,
+    );
 
     await checkoutEnvironment(projRoot, { envName: firstEnvName });
     await removeEnvironment(projRoot, { envName: secondEnvName });
@@ -82,6 +99,7 @@ describe('upload and delete parameters', () => {
     );
 
     await deleteProject(projRoot);
+    deletionRequired = false;
     await expectParametersOptionalValue(
       [],
       expectedParamsAfterAddFunc.map((pair) => pair.name),