+++ packages/amplify-app/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.8](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-app@5.0.6...@aws-amplify/amplify-app@5.0.8) (2023-05-25)
+
+
+### Bug Fixes
+
+* version bump ([f820824](https://github.com/aws-amplify/amplify-cli/commit/f82082416187bd1dd33de9f7b35753026ac17eea))
+
+
+
+
+
+## [5.0.7](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-app@5.0.5...@aws-amplify/amplify-app@5.0.7) (2023-05-25)
+
+**Note:** Version bump only for package @aws-amplify/amplify-app
+
+
+
+
+
+## [5.0.6](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-app@5.0.5...@aws-amplify/amplify-app@5.0.6) (2023-05-25)
+
+**Note:** Version bump only for package @aws-amplify/amplify-app
+
+
+
+
+
 ## [5.0.5](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-app@5.0.4...@aws-amplify/amplify-app@5.0.5) (2023-05-11)
 
 **Note:** Version bump only for package @aws-amplify/amplify-app